Add support for controller input

+++ Framework/Source/Graphics/Text.h
@@ -18,7 +18,7 @@ namespace Framework
 		void setColor(float r, float g, float b, float a = 1);
 		int getFontHeight() const;
 		void useFontTexture() const;
-		StringStream operator()(float x, float y);
+		StringStream operator()(int x, int y);
 
 	private:
 		typedef std::unordered_map<string, std::shared_ptr<Font>> Fonts;
@@ -31,7 +31,7 @@ namespace Framework
 		struct DrawConfiguration
 		{
 			Text* text;
-			float x, y;
+			int x, y;
 		};
 
 		static const int MAX_STRING_LENGTH = 1 << 10;
@@ -50,7 +50,7 @@ namespace Framework
 		std::shared_ptr<Font> findFont(const string& family) const;
 		void initialize(int windowWidth, int windowHeight);
 		void windowResizedEventHandler(int width, int height);
-		void draw(float x, float y, const std::u32string& text);
+		void draw(int x, int y, const std::u32string& text);
 		void applyWindowSize();
 
 		static void drawFromStreamCallback(const StringStream& stream, void* data);
